% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{YonseiGLP1-package}
\alias{YonseiGLP1}
\alias{YonseiGLP1-package}
\title{YonseiGLP1: A Package Skeleton for Comparative Effectiveness Studies}
\description{
A skeleton package, to be used as a starting point when implementing comparative effect studies.
}
\keyword{internal}
